// frontend/theme/types.ts
export interface AppColors {
    background: string
    foreground: string
    card: string
    textPrimary: string
    textSecondary: string
    accentRed: string
    accentGreen: string
    topBar: string
    navigationActive: string
    navigationInactive: string
}

export type ThemeName = "dark" | "light"